+++ R/quantile_data.R
@@ -1,0 +1,26 @@
+#' @title Quantile data for quantile_plot example.
+#' @description Quantile data for quantile_plot example.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: include Value, Group columns.
+#' @usage data(quantile_data)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/QuantileQuantile/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(quantile_data)
+#'
+#' # 3. View example data
+#' quantile_data
+#'
+"quantile_data"
+
+# quantile_data <- data
+# usethis::use_data(quantile_data)